Separate fake asset tracker service from real
