+++ il_ws/lib/il_fake_services.dart
@@ -1,0 +1,3 @@
+library fake_services;
+
+export 'src/misc/fake_services/fake_asset_location_tracker.dart';
